+++ clients/admin-ui/src/features/privacy-requests/dashboard/hooks/usePrivacyRequestBulkActions.test.tsx
@@ -68,7 +68,6 @@ describe("usePrivacyRequestBulkActions", () => {
     pendingRequest1,
     completeRequest,
   ];
-  const mockClearSelectedIds = jest.fn();
 
   beforeEach(() => {
     jest.clearAllMocks();
@@ -80,7 +79,6 @@ describe("usePrivacyRequestBulkActions", () => {
       usePrivacyRequestBulkActions({
         requests: mockRequests,
         selectedIds: ["1", "3"],
-        clearSelectedIds: mockClearSelectedIds,
         modalApi: mockModalApi,
       }),
     );
@@ -119,7 +117,6 @@ describe("usePrivacyRequestBulkActions", () => {
         usePrivacyRequestBulkActions({
           requests: [pendingRequest1, pendingRequest2],
           selectedIds: ["1", "2"],
-          clearSelectedIds: mockClearSelectedIds,
           modalApi: mockModalApi,
         }),
       );
@@ -163,7 +160,6 @@ describe("usePrivacyRequestBulkActions", () => {
         usePrivacyRequestBulkActions({
           requests: [pendingRequest1, pendingRequest2],
           selectedIds: ["1", "2"],
-          clearSelectedIds: mockClearSelectedIds,
           modalApi: mockModalApi,
         }),
       );
@@ -209,7 +205,6 @@ describe("usePrivacyRequestBulkActions", () => {
         usePrivacyRequestBulkActions({
           requests: [pendingRequest1, completeRequest],
           selectedIds: ["1", "3"],
-          clearSelectedIds: mockClearSelectedIds,
           modalApi: mockModalApi,
         }),
       );
@@ -244,7 +239,6 @@ describe("usePrivacyRequestBulkActions", () => {
         usePrivacyRequestBulkActions({
           requests: [pendingRequest1, completeRequest],
           selectedIds: ["1", "3"],
-          clearSelectedIds: mockClearSelectedIds,
           modalApi: mockModalApi,
         }),
       );
@@ -280,7 +274,6 @@ describe("usePrivacyRequestBulkActions", () => {
         usePrivacyRequestBulkActions({
           requests: [pendingRequest1],
           selectedIds: ["1"],
-          clearSelectedIds: mockClearSelectedIds,
           modalApi: mockModalApi,
         }),
       );
@@ -305,7 +298,6 @@ describe("usePrivacyRequestBulkActions", () => {
         usePrivacyRequestBulkActions({
           requests: [pendingRequest1],
           selectedIds: ["1"],
-          clearSelectedIds: mockClearSelectedIds,
           modalApi: mockModalApi,
         }),
       );