+++ clients/admin-ui/src/features/privacy-requests/dashboard/hooks/usePrivacyRequestBulkActions.tsx
@@ -4,7 +4,7 @@ import {
   Icons,
   useMessage,
 } from "fidesui";
-import { useCallback, useEffect, useMemo } from "react";
+import { useCallback, useMemo } from "react";
 
 import { pluralize } from "~/features/common/utils";
 import { PrivacyRequestResponse } from "~/types/api";
@@ -26,7 +26,6 @@ type ModalInstance = ReturnType<typeof modal.useModal>[0];
 interface UsePrivacyRequestBulkActionsProps {
   requests: PrivacyRequestResponse[];
   selectedIds: React.Key[];
-  clearSelectedIds: () => void;
   modalApi: ModalInstance;
 }
 
@@ -88,7 +87,6 @@ const formatResultMessage = (
 export const usePrivacyRequestBulkActions = ({
   requests,
   selectedIds,
-  clearSelectedIds,
   modalApi,
 }: UsePrivacyRequestBulkActionsProps) => {
   const selectedRequests = useMemo(
@@ -97,11 +95,6 @@ export const usePrivacyRequestBulkActions = ({
   );
 
   const messageApi = useMessage();
-
-  // Clear selected requests when the data changes. eg. with pagination, filters or actions performed
-  useEffect(() => {
-    clearSelectedIds();
-  }, [requests, clearSelectedIds]);
 
   // Mutation hooks for the actions
   const [bulkApproveRequest] = useBulkApproveRequestMutation();